tsに拡張子変更
