+++ .gitignore
@@ -1,3 +1,4 @@
 front/node_modules
 .env
 .env.local
+dist
