+++ front/index.html
@@ -6,6 +6,6 @@
   </head>
   <body>
     <div id="content"></div>
-    <script type="module" src="/src/main.jsx"></script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
